feat camc: add comments and move base logic to another file
